Move MGS classes to a module
